Add documentation to backup files
